requirements.txt
# -----------------------------
# Core Python Libraries
# -----------------------------
numpy==1.24.4          # Numerical computations
pandas==2.0.3          # Data handling & manipulation
matplotlib==3.7.2      # Plotting graphs
seaborn==0.12.2        # Advanced visualizations

# -----------------------------
# Deep Learning Framework
# -----------------------------
tensorflow==2.13.0      # Deep learning (CNN, Transfer Learning, Grad-CAM)
keras==2.13.1           # High-level API for TensorFlow models

# -----------------------------
# Machine Learning Utilities
# -----------------------------
scikit-learn==1.3.0     # Evaluation metrics, confusion matrix, ROC, etc.
joblib==1.3.2           # Saving & loading trained models
h5py==3.9.0             # Saving models in HDF5 format

# -----------------------------
# Explainability Libraries
# -----------------------------
shap==0.44.1            # SHAP explainability for model interpretation
opencv-python==4.8.0.76 # Image preprocessing (cv2)

# -----------------------------
# Web App Frontend
# -----------------------------
streamlit==1.25.0       # Streamlit for frontend UI
Pillow==10.0.0          # Image handling (PIL)

# -----------------------------
# Other Utilities
# -----------------------------
tqdm==4.65.0            # Progress bars during training
protobuf==4.23.4        # Required for TensorFlow compatibility
